% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_apply.R
\name{iso_apply}
\alias{iso_apply}
\title{sensible row- or column- wise operations against a DelayedMatrix}
\usage{
iso_apply(X, MARGIN, FUN, by = NULL)
}
\arguments{
\item{X}{a DelayedArray (e.g. a DelayedMatrix backed by HDF5Array)}

\item{MARGIN}{apply across rows (1) or columns (2), as in base::apply}

\item{FUN}{the function to apply to each row or column (see above)}

\item{by}{the number of rows or columns to load at one time}
}
\value{
\preformatted{   a new DelayedArray (presumably of the same dimensions as X)
}
}
\description{
For good performance, entire chunks should be loaded; consequently,
if MARGIN == 1 (rows), \code{by} should be a multiple of chunkdim(X)\link{1}, and
if MARGIN == 2 (columns), \code{by} should be a multiple of chunkdim(X)\link{2}.
(There is a check included for this, but the principle is worth noting.)
}
\details{
Remember to setRealizationBackend("HDF5Array") or TileDBArray or whatever.
}
