% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCS.R
\name{LCS}
\alias{LCS}
\title{longest common (gapped) substring for a sequence of strings}
\usage{
LCS(..., sep = "")
}
\arguments{
\item{...}{the strings}

\item{sep}{their internal separator (default is "")}
}
\value{
\preformatted{ their LCS
}
}
\description{
This is built for comfort, not for speed.
Don't use it to build phylogenetic trees.
}
