% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeScNMT.R
\name{mergeScNMT}
\alias{mergeScNMT}
\title{merge a bunch of scNMT raw methylation files (chrom, start, \%meth format)}
\usage{
mergeScNMT(
  tsvs,
  gen = "GRCm38",
  loci = NULL,
  saveGR = TRUE,
  saveSE = FALSE,
  HDF5 = FALSE,
  dir = "scNMT_meth",
  BPPARAM = SerialParam()
)
}
\arguments{
\item{tsvs}{scNMT raw methylation files}

\item{gen}{what genome they correspond to (default is GRCm38)}

\item{loci}{a pre-merged GRanges with the union of all tsvs' ranges (NULL)}

\item{saveGR}{save the intermediate \code{loci} object, if loci=NULL? (TRUE)}

\item{saveSE}{save the intermediate \code{se} object before filling it? (FALSE)}

\item{HDF5}{use an HDF5-backed SummarizedExperiment for loading? (FALSE)}

\item{dir}{if using HDF5 and/or saving loci, path to use ("scNMT_meth")}

\item{BPPARAM}{BiocParallelParam object (default BiocParallel::SerialParam())}
}
\value{
\preformatted{   a SummarizedExperiment, perhaps HDF5-backed, with merged data
}
}
\description{
NOTE: see bsseq:::.constructCounts for some background on doing this in HDF5
}
