% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanScNMT.R
\name{scanScNMT}
\alias{scanScNMT}
\title{import an scNMT raw methylation file (chrom, start, \%meth)}
\usage{
scanScNMT(
  tsv,
  dry = FALSE,
  gen = "GRCm38",
  seqinf = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{tsv}{a tsv or tsv.gz filename (will figure out what to do with it)}

\item{dry}{dry run? (just check that the file is tabixed?) (FALSE)}

\item{gen}{what genome the file corresponds to (default is GRCm38)}

\item{seqinf}{add seqInfo? (leave FALSE if running behind a firewall)}

\item{verbose}{be verbose? (FALSE)}

\item{...}{additional arguments to pass on to rtracklayer::import}
}
\value{
\preformatted{   a GRanges with appropriate seqinfo and scores turned to betas
}
}
\description{
import an scNMT raw methylation file (chrom, start, \%meth)
}
