% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMTFileListToGR.R
\name{scNMTFileListToGR}
\alias{scNMTFileListToGR}
\title{somewhat of an evolved version of tabixToGR with checkpointing}
\usage{
scNMTFileListToGR(tfl, BPPARAM = SerialParam(), verbose = TRUE)
}
\arguments{
\item{tfl}{a TabixFileList or something that can become one}

\item{BPPARAM}{BiocParallel parameters (default is SerialParam())}

\item{verbose}{be verbose? (TRUE)}
}
\value{
\preformatted{     a GRanges
}
}
\description{
somewhat of an evolved version of tabixToGR with checkpointing
}
