% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeScNMT.R
\name{mergeScNMT}
\alias{mergeScNMT}
\title{merge a bunch of scNMT raw (met|acc) files (chrom, start, \%meth/\%acc format)}
\usage{
mergeScNMT(
  tsvs,
  gen = "GRCm38",
  loci = NULL,
  saveGR = TRUE,
  saveSE = FALSE,
  HDF5 = FALSE,
  what = c("meth", "acc"),
  dir = "scNMT",
  seqinf = TRUE,
  verbose = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{tsvs}{scNMT raw CpG (methylation) or GpC (accessibility) files}

\item{gen}{what genome they correspond to (default is GRCm38)}

\item{loci}{a pre-merged GRanges with the union of all tsvs' ranges (NULL)}

\item{saveGR}{save the intermediate \code{loci} object, if loci=NULL? (TRUE)}

\item{saveSE}{save the intermediate \code{se} object before filling it? (FALSE)}

\item{HDF5}{use an HDF5-backed SummarizedExperiment for loading? (FALSE)}

\item{what}{methylation ("meth") or accessibility ("acc") data? ("meth")}

\item{dir}{if using HDF5 and/or saving loci, path to use ("scNMT_"what)}

\item{seqinf}{add seqInfo? (TRUE; set FALSE if running behind a firewall)}

\item{verbose}{be extra verbose? (FALSE)}

\item{BPPARAM}{BiocParallelParam object (default BiocParallel::SerialParam())}
}
\value{
\preformatted{   a SummarizedExperiment, perhaps HDF5-backed, with merged data
}
}
\description{
NOTE: see bsseq:::.constructCounts for some background on doing this in HDF5
}
