% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMTtoGR.R
\name{scNMTtoGR}
\alias{scNMTtoGR}
\title{take an scNMT CpG methylation file and read its coordinates into a GRanges}
\usage{
scNMTtoGR(filename, verbose = TRUE, is0based = FALSE, which = NULL, ...)
}
\arguments{
\item{filename}{the file}

\item{verbose}{squawk? (TRUE)}

\item{is0based}{is it a UCSC file? (FALSE)}

\item{which}{Optional instance of 'GRanges' or 'IntegerRangesList' (NULL)}

\item{...}{Optional arguments, currently ignored}
}
\value{
\preformatted{     a GRanges with the tabix-derived coordinates, but no $score
}
}
\description{
This function is primarily to Reduce(union, aBunchOfGRs) for aggregation
}
